requirements.txt
## The following requirements were added by pip freeze:
blinker==1.6.2
certifi==2023.5.7
click==8.1.6
Flask==2.3.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
python-dotenv==1.0.0
PyYAML==6.0.1
sentry-sdk==1.28.1
urllib3==2.0.4
Werkzeug==2.3.6
